+++ CHANGELOG.md
@@ -3,12 +3,17 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [5.0.8] - 2022-08-29
+## [5.0.9] - 2022-08-31
 ### Added
 - Options to variables tab for CTSS API.
 ### Fixed
 - Updating the position of the sticky table header when changing the content of
   the header.
+
+## [5.0.8] - 2022-08-30
+### Changed
+- Replace mat-autocomplete with autocomplete-lhc controls.
+- Match list value on search parameter value autocomplete controls.
 
 ## [5.0.7] - 2022-08-29
 ### Added